Several cleanups
- added testing
- removed the out of bounds error in ErrorCode
- minimized unwraps as much as possible
- minimized primitive type conversions as much as possible
- updated ata_pio to try to read as much as possible
- updated diskstreamer to handle hard disks that didn't read enough
- updated the makefile to support testing, debug, and release modes

+++ src/disk/diskreader.rs
@@ -7,12 +7,12 @@ use super::ata_pio::AtaPio;
 pub trait DiskReader: Send + Sync {
     fn read(&self, lba: usize, out: &mut [u8], total: usize) -> Result<usize, ErrorCode>;
     fn write(&self, lba: usize, data: &mut [u8]) -> Result<(), ErrorCode>;
-    fn resolve(index: usize) -> Result<Self, ErrorCode>
+    fn resolve(index: u32) -> Result<Self, ErrorCode>
     where
         Self: Sized;
 }
 
-pub fn find_diskreader(disk_id: usize) -> Result<Box<dyn DiskReader>, ErrorCode> {
+pub fn find_diskreader(disk_id: u32) -> Result<Box<dyn DiskReader>, ErrorCode> {
     match AtaPio::resolve(disk_id) {
         Ok(val) => return Ok(Box::new(val)),
         Err(ErrorCode::DiskNotUs) => (),
